// frontend/types.ts
export interface Image {
  id: number;
  image_url: string;
  alt_text: string;
  caption?: string;
  show_on_home?: boolean;
  image_type: string;
  linked_object_type: string | null;
}